+++ tuxemon/states/start/__init__.py
@@ -19,7 +19,7 @@ from tuxemon.platform.const import buttons
 from tuxemon.platform.events import PlayerInput
 from tuxemon.save import get_index_of_latest_save
 from tuxemon.session import local_session
-from tuxemon.state import State
+from tuxemon.state.state import State
 from tuxemon.time_handler import today_ordinal
 
 logger = logging.getLogger(__name__)
