+++ tuxemon/states/transition/flash.py
@@ -11,7 +11,7 @@ from tuxemon import prepare
 from tuxemon.graphics import ColorLike
 from tuxemon.platform.events import PlayerInput
 from tuxemon.rumble.tools import RumbleParams
-from tuxemon.state import State
+from tuxemon.state.state import State
 
 logger = logging.getLogger(__name__)
