+++ tuxemon/states/transition/pixelation.py
@@ -9,7 +9,7 @@ from pygame.surface import Surface
 from pygame.transform import scale
 
 from tuxemon.platform.events import PlayerInput
-from tuxemon.state import State
+from tuxemon.state.state import State
 
 logger = logging.getLogger(__name__)
