+++ tuxemon/states/transition/swirl.py
@@ -10,7 +10,7 @@ from pygame.transform import rotate, scale
 
 from tuxemon import prepare
 from tuxemon.platform.events import PlayerInput
-from tuxemon.state import State
+from tuxemon.state.state import State
 
 logger = logging.getLogger(__name__)
